@@ -1,17 +1,19 @@
 name: fwfh_webview
-version: 0.6.2+2
+version: 0.7.0+1
 description: WidgetFactory extension to render IFRAME with the official WebView plugin.
 homepage: https://github.com/daohoangson/flutter_widget_from_html
 
 environment:
-  flutter: ">=2.5.0"
-  sdk: ">=2.14.0 <3.0.0"
+  flutter: ">=3.0.0"
+  sdk: ">=2.17.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_widget_from_html_core: ">=0.6.1 <0.9.0"
-  webview_flutter: ">=2.2.0 <4.0.0"
+  flutter_widget_from_html_core: ">=0.8.0 <0.11.0"
+  webview_flutter: ^4.0.1
+  webview_flutter_android: ^3.0.0
+  webview_flutter_wkwebview: ^3.0.0
 
 dependency_overrides:
   flutter_widget_from_html_core:
@@ -24,3 +26,5 @@ dev_dependencies:
     sdk: flutter
   lint: any
   measurer: ^2.1.1
+  plugin_platform_interface: any
+  webview_flutter_platform_interface: ^2.0.0
