@@ -1,18 +1,20 @@
 import 'package:demo_app/screens/hello_world.dart' as enhanced;
 import 'package:demo_app/widgets/popup_menu.dart';
+import 'package:demo_app/widgets/selection_area.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_widget_from_html_core/flutter_widget_from_html_core.dart';
 
 class HelloWorldCoreScreen extends StatelessWidget {
-  const HelloWorldCoreScreen({Key key}) : super(key: key);
+  const HelloWorldCoreScreen({super.key});
 
   @override
-  Widget build(BuildContext context) => Scaffold(
+  Widget build(BuildContext context) => SelectionAreaScaffold(
         appBar: AppBar(
           title: const Text('HelloWorldCoreScreen'),
           actions: const [
             PopupMenu(
               scrollToTop: true,
+              toggleIsSelectable: true,
             ),
           ],
         ),
