@@ -367,14 +367,21 @@ ${htmlDesc * 3}
 final globalKey = GlobalKey<HtmlWidgetState>();
 
 Future<void> pumpWidget(WidgetTester tester, Widget child) async {
+  // TODO: remove lint ignore when our minimum Flutter version >= 3.10
+  // ignore: deprecated_member_use
   tester.binding.window.physicalSizeTestValue = const Size(200, 200);
+  // ignore: deprecated_member_use
   addTearDown(tester.binding.window.clearPhysicalSizeTestValue);
-
+  // ignore: deprecated_member_use
   tester.binding.window.devicePixelRatioTestValue = 1.0;
+  // ignore: deprecated_member_use
   addTearDown(tester.binding.window.clearDevicePixelRatioTestValue);
-
-  tester.binding.window.textScaleFactorTestValue = 1.0;
-  addTearDown(tester.binding.window.clearTextScaleFactorTestValue);
+  // ignore: deprecated_member_use
+  tester.binding.window.platformDispatcher.textScaleFactorTestValue = 1.0;
+  addTearDown(
+    // ignore: deprecated_member_use
+    tester.binding.window.platformDispatcher.clearTextScaleFactorTestValue,
+  );
 
   await tester.pumpWidget(MaterialApp(home: child));
   await tester.pump();
