@@ -87,7 +87,7 @@ void main() {
           ),
         );
 
-    void _expect(Widget? built1, Widget? built2, Matcher matcher) {
+    void enableCachingExpect(Widget? built1, Widget? built2, Matcher matcher) {
       final widget1 = (built1! as TshWidget).child;
       final widget2 = (built2! as TshWidget).child;
       expect(widget1 == widget2, matcher);
@@ -101,7 +101,7 @@ void main() {
 
       await explain(tester, html, enableCaching: true);
       final built2 = helper.buildCurrentState();
-      _expect(built1, built2, isTrue);
+      enableCachingExpect(built1, built2, isTrue);
     });
 
     testWidgets('rebuild new html', (WidgetTester tester) async {
@@ -125,11 +125,11 @@ void main() {
       await explain(
         tester,
         html,
-        baseUrl: Uri.http('domain.com', ''),
+        baseUrl: Uri.http('domain.com'),
         enableCaching: true,
       );
       final built2 = helper.buildCurrentState();
-      _expect(built1, built2, isFalse);
+      enableCachingExpect(built1, built2, isFalse);
     });
 
     testWidgets('rebuild new buildAsync', (tester) async {
@@ -141,7 +141,7 @@ void main() {
 
       await explain(tester, html, buildAsync: false, enableCaching: true);
       final built2 = helper.buildCurrentState();
-      _expect(built1, built2, isFalse);
+      enableCachingExpect(built1, built2, isFalse);
     });
 
     testWidgets('rebuild new enableCaching', (tester) async {
@@ -153,7 +153,7 @@ void main() {
 
       await explain(tester, html, enableCaching: false);
       final built2 = helper.buildCurrentState();
-      _expect(built1, built2, isFalse);
+      enableCachingExpect(built1, built2, isFalse);
     });
 
     testWidgets('rebuild new rebuildTriggers', (tester) async {
@@ -175,7 +175,7 @@ void main() {
         rebuildTriggers: [2],
       );
       final built2 = helper.buildCurrentState();
-      _expect(built1, built2, isFalse);
+      enableCachingExpect(built1, built2, isFalse);
     });
 
     testWidgets('rebuild new textStyle', (tester) async {
@@ -201,7 +201,7 @@ void main() {
 
       await explain(tester, html, enableCaching: false);
       final built2 = helper.buildCurrentState();
-      _expect(built1, built2, isFalse);
+      enableCachingExpect(built1, built2, isFalse);
     });
   });
 
@@ -297,33 +297,6 @@ void main() {
         ),
       );
       expect(e, equals('[Column:children=[RichText:(:Foo)],[Text:Bar]]'));
-    });
-  });
-
-  group('isSelectable', () {
-    const html = 'Foo';
-
-    testWidgets('renders RichText', (WidgetTester tester) async {
-      final hw = HtmlWidget(html, key: helper.hwKey);
-      final explained = await explain(tester, hw);
-      expect(explained, equals('[RichText:(:Foo)]'));
-    });
-
-    testWidgets('renders SelectableText', (WidgetTester tester) async {
-      final hw = HtmlWidget(html, isSelectable: true, key: helper.hwKey);
-      final explained = await explain(tester, hw);
-      expect(explained, equals('[SelectableText:(:Foo)]'));
-    });
-
-    testWidgets('renders onSelectionChanged', (WidgetTester tester) async {
-      final hw = HtmlWidget(
-        html,
-        isSelectable: true,
-        key: helper.hwKey,
-        onSelectionChanged: (_, __) {},
-      );
-      final explained = await explain(tester, hw);
-      expect(explained, equals('[SelectableText:+onSelectionChanged,(:Foo)]'));
     });
   });
 
