@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:fwfh_webview/src/web_view/web_view.dart';
 import 'package:measurer/measurer.dart';
+import 'package:webview_flutter/webview_flutter.dart';
 
 import 'mock_webview_platform.dart';
 
@@ -13,21 +14,19 @@ void main() {
     const url = 'http://domain.com/?document.body.scrollWidth=1000&'
         'document.body.scrollHeight=1000';
 
-    var _aspectRatio = double.nan;
+    var aspectRatio = double.nan;
 
     // ignore: prefer_function_declarations_over_variables
     final run = (
       WidgetTester tester, {
       String urlQueryParams = '',
-      bool unsupportedWorkaroundForIssue375 = false,
       Widget Function(Widget)? wrapper,
     }) async {
       final child = Measurer(
-        onMeasure: (v, _) => _aspectRatio = v.width / v.height,
+        onMeasure: (v, _) => aspectRatio = v.width / v.height,
         child: WebView(
           '$url&$urlQueryParams',
           aspectRatio: defaultAspectRatio,
-          unsupportedWorkaroundForIssue375: unsupportedWorkaroundForIssue375,
         ),
       );
 
@@ -47,7 +46,7 @@ void main() {
 
     // ignore: prefer_function_declarations_over_variables
     final expectAspectRatioEquals = (double expected) => expect(
-          _aspectRatio.toStringAsFixed(2),
+          aspectRatio.toStringAsFixed(2),
           equals(expected.toStringAsFixed(2)),
         );
 
@@ -57,73 +56,16 @@ void main() {
       await tester.pumpAndSettle();
     };
 
-    testWidgets('unsupportedWorkaroundForIssue375=false', (tester) async {
+    testWidgets('ratio 1.0', (tester) async {
       await run(tester);
       expectAspectRatioEquals(1.0);
       await cleanUp(tester);
     });
 
-    group('unsupportedWorkaroundForIssue375=true', () {
-      testWidgets('requires too much height', (WidgetTester tester) async {
-        const testSize = Size(800, 600);
-        tester.binding.window.physicalSizeTestValue = testSize;
-        addTearDown(tester.binding.window.clearPhysicalSizeTestValue);
-
-        await run(tester, unsupportedWorkaroundForIssue375: true);
-        expectAspectRatioEquals(testSize.width / testSize.height);
-
-        await cleanUp(tester);
-      });
-
-      testWidgets('height fits', (WidgetTester tester) async {
-        const testSize = Size(600, 800);
-        tester.binding.window.physicalSizeTestValue = testSize;
-        addTearDown(tester.binding.window.clearPhysicalSizeTestValue);
-
-        await run(tester, unsupportedWorkaroundForIssue375: true);
-        expectAspectRatioEquals(1.0);
-
-        await cleanUp(tester);
-      });
-
-      testWidgets('in vertical ListView', (WidgetTester tester) async {
-        const testSize = Size(800, 600);
-        tester.binding.window.physicalSizeTestValue = testSize;
-        addTearDown(tester.binding.window.clearPhysicalSizeTestValue);
-
-        await run(
-          tester,
-          unsupportedWorkaroundForIssue375: true,
-          wrapper: (child) => ListView(children: [child]),
-        );
-        expectAspectRatioEquals(testSize.width / testSize.height);
-
-        await cleanUp(tester);
-      });
-
-      testWidgets('in horizontal ListView', (WidgetTester tester) async {
-        const testSize = Size(800, 600);
-        tester.binding.window.physicalSizeTestValue = testSize;
-        addTearDown(tester.binding.window.clearPhysicalSizeTestValue);
-
-        await run(
-          tester,
-          unsupportedWorkaroundForIssue375: true,
-          wrapper: (child) => ListView(
-            scrollDirection: Axis.horizontal,
-            children: [child],
-          ),
-        );
-        expectAspectRatioEquals(testSize.width / testSize.height);
-
-        await cleanUp(tester);
-      });
-    });
-
     testWidgets('handles js error', (tester) async {
       await run(
         tester,
-        urlQueryParams: 'runJavascriptReturningResult=error',
+        urlQueryParams: 'runJavaScriptReturningResult=error',
       );
       expectAspectRatioEquals(defaultAspectRatio);
       await cleanUp(tester);
@@ -153,11 +95,11 @@ void main() {
       await tester.pumpAndSettle();
 
       const url2 = 'http://domain.com/2';
-      final result2 = await FakeWebViewController.instance?.handler
-          .onNavigationRequest(url: url2, isForMainFrame: true);
+      final result2 = await FakeWebViewController.instance
+          ?.onNavigationRequest(url: url2, isMainFrame: true);
 
       expect(navigationRequestUrls, equals([url2]));
-      expect(result2, isFalse);
+      expect(result2, equals(NavigationDecision.prevent));
     });
 
     testWidgets('skips callback (initial url)', (WidgetTester tester) async {
@@ -181,11 +123,11 @@ void main() {
       );
       await tester.pumpAndSettle();
 
-      final result = await FakeWebViewController.instance?.handler
-          .onNavigationRequest(url: url, isForMainFrame: true);
+      final result = await FakeWebViewController.instance
+          ?.onNavigationRequest(url: url, isMainFrame: true);
 
       expect(navigationRequestUrls, equals([]));
-      expect(result, isTrue);
+      expect(result, equals(NavigationDecision.navigate));
     });
 
     testWidgets('skips callback (first url)', (WidgetTester tester) async {
@@ -210,11 +152,11 @@ void main() {
       );
       await tester.pumpAndSettle();
 
-      final result2 = await FakeWebViewController.instance?.handler
-          .onNavigationRequest(url: url2, isForMainFrame: true);
+      final result2 = await FakeWebViewController.instance
+          ?.onNavigationRequest(url: url2, isMainFrame: true);
 
       expect(navigationRequestUrls, equals([]));
-      expect(result2, isTrue);
+      expect(result2, equals(NavigationDecision.navigate));
     });
 
     testWidgets('skips callback (not main frame)', (WidgetTester tester) async {
@@ -239,11 +181,11 @@ void main() {
       await tester.pumpAndSettle();
 
       const url2 = 'http://domain.com/2';
-      final result2 = await FakeWebViewController.instance?.handler
-          .onNavigationRequest(url: url2, isForMainFrame: false);
+      final result2 = await FakeWebViewController.instance
+          ?.onNavigationRequest(url: url2, isMainFrame: false);
 
       expect(navigationRequestUrls, equals([]));
-      expect(result2, isTrue);
+      expect(result2, equals(NavigationDecision.navigate));
     });
   });
 
